+++ mobile/android/geckoview/src/androidTest/assets/web_extensions/test-support/test-api.js
@@ -108,6 +108,10 @@ this.test = class extends ExtensionAPI {
         async addHistogram(id, value) {
           return Services.telemetry.getHistogramById(id).add(value);
         },
+
+        async setScalar(id, value) {
+          return Services.telemetry.scalarSet(id, value);
+        },
       },
     };
   }