+++ mobile/android/geckoview/src/androidTest/java/org/mozilla/geckoview/test/util/RuntimeCreator.java
@@ -35,9 +35,30 @@ public class RuntimeCreator {
         public RuntimeTelemetry.Delegate delegate = null;
 
         @Override
-        public void onTelemetryReceived(@NonNull RuntimeTelemetry.Metric metric) {
+        public void onHistogram(@NonNull RuntimeTelemetry.Metric<long[]> metric) {
             if (delegate != null) {
-                delegate.onTelemetryReceived(metric);
+                delegate.onHistogram(metric);
+            }
+        }
+
+        @Override
+        public void onBooleanScalar(@NonNull RuntimeTelemetry.Metric<Boolean> metric) {
+            if (delegate != null) {
+                delegate.onBooleanScalar(metric);
+            }
+        }
+
+        @Override
+        public void onStringScalar(@NonNull RuntimeTelemetry.Metric<String> metric) {
+            if (delegate != null) {
+                delegate.onStringScalar(metric);
+            }
+        }
+
+        @Override
+        public void onLongScalar(@NonNull RuntimeTelemetry.Metric<Long> metric) {
+            if (delegate != null) {
+                delegate.onLongScalar(metric);
             }
         }
     }