+++ widget/android/GeckoTelemetryDelegate.h
@@ -48,23 +48,35 @@ class GeckoTelemetryDelegate final
       samples->AppendElement(static_cast<int64_t>(aSamples[i]));
     }
 
-    mProxy->DispatchTelemetry(
+    mProxy->DispatchHistogram(
         aName,
         mozilla::jni::LongArray::New(samples->Elements(), samples->Length()));
   }
 
   void ReceiveBoolScalarValue(const nsCString& aName, bool aValue) override {
-    MOZ_ASSERT_UNREACHABLE("ReceiveBoolScalarValue unimplemented");
+    if (!mozilla::jni::IsAvailable() || !mProxy) {
+      return;
+    }
+
+    mProxy->DispatchBooleanScalar(aName, aValue);
   }
 
   void ReceiveStringScalarValue(const nsCString& aName,
                                 const nsCString& aValue) override {
-    MOZ_ASSERT_UNREACHABLE("ReceiveStringScalarValue unimplemented");
+    if (!mozilla::jni::IsAvailable() || !mProxy) {
+      return;
+    }
+
+    mProxy->DispatchStringScalar(aName, aValue);
   }
 
   void ReceiveUintScalarValue(const nsCString& aName,
                               uint32_t aValue) override {
-    MOZ_ASSERT_UNREACHABLE("ReceiveUintScalarValue unimplemented");
+    if (!mozilla::jni::IsAvailable() || !mProxy) {
+      return;
+    }
+
+    mProxy->DispatchLongScalar(aName, static_cast<int64_t>(aValue));
   }
 
   mozilla::java::RuntimeTelemetry::Proxy::GlobalRef mProxy;